requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
pluggy-sdk>=1.0.0
python-dotenv>=1.0.0
